Updated builddist.rb to package the Embree DLL (AuxFunctions.dll) Updated VS2008 projects for Embree libraries (all configs not including the .lib, renamed DLL) Fixed size_t warnings
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2910 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/Transmungify.cpp
@@ -37,7 +37,7 @@ static inline unsigned char toUChar(uint32 x)
 
 bool Transmungify::encrypt(const std::string& src_string, std::vector<uint32>& dst_dwords)
 {
-	return encrypt(src_string.c_str(), src_string.size(), dst_dwords);
+	return encrypt(src_string.c_str(), (uint32)src_string.size(), dst_dwords);
 }
